Move the AppImage release scripts under scripts/appimage
They only serve the AppImage target, not deb, rpm or any other platform, and
a folder says so where a file prefix did not.
