+++ scripts/appimage/elf.test.ts
