+++ scripts/appimage/elf.ts
